third commit

+++ server/controller/uploadCtrl.js
@@ -2,6 +2,7 @@ const PopImages = require("../models/populerModel");
 const LogoImages = require("../models/logoModel");
 const SocialImages = require("../models/socialModel");
 const BrandImages = require("../models/brandModel");
+const BrandDImages = require("../models/brandDModel");
 const { uploadImg } = require("../utils/cloudinary");
 const fs = require("fs");
 
@@ -87,6 +88,29 @@ exports.uploadBrand = async (req, res, next) => {
     for (const file of files) {
       const { secure_url, public_id } = await uploader(file.path);
       const newImg = await BrandImages.create({ public_id, secure_url });
+      urls.push(newImg);
+      fs.unlinkSync(file.path);
+    }
+
+    res.status(201).json(urls);
+  } catch (error) {
+    next(error);
+  }
+};
+
+exports.uploadBrandD = async (req, res, next) => {
+  const files = req.files;
+  try {
+    if (!files) {
+      const error = new Error("Theres is no images to upload");
+      error.statusCode = 403;
+      throw error;
+    }
+    const uploader = (file) => uploadImg(file, "brand-design");
+    const urls = [];
+    for (const file of files) {
+      const { secure_url, public_id } = await uploader(file.path);
+      const newImg = await BrandDImages.create({ public_id, secure_url });
       urls.push(newImg);
       fs.unlinkSync(file.path);
     }
